# Add some micro optimizations