--- a/src/log_file.rs
+++ b/src/log_file.rs
@@ -10,7 +10,7 @@ use nix::errno::Errno;
 use nix::fcntl::{self, OFlag};
 use nix::sys::stat::Mode;
 use nix::sys::uio::{pread, pwrite};
-use nix::unistd::{close, fdatasync, ftruncate, lseek, Whence};
+use nix::unistd::{close, ftruncate, lseek, Whence};
 use nix::NixPath;
 
 /// A `LogFd` is a RAII file that provides basic I/O functionality.
@@ -59,7 +59,14 @@ impl LogFd {
     }
 
     pub fn sync(&self) -> IoResult<()> {
-        fdatasync(self.0).map_err(|e| from_nix_error(e, "fsync"))
+        #[cfg(target_os = "linux")]
+        {
+            nix::unistd::fdatasync(self.0).map_err(|e| from_nix_error(e, "fdatasync"))
+        }
+        #[cfg(not(target_os = "linux"))]
+        {
+            nix::unistd::fsync(self.0).map_err(|e| from_nix_error(e, "fsync"))
+        }
     }
 
     pub fn read(&self, mut offset: usize, buf: &mut [u8]) -> IoResult<usize> {